Move AppBtn components into its own subfolder for organisation and update imports in NavigationBar component
